Third-EYE-AI-SOC/requirements.txt
fastapi
uvicorn
requests
streamlit
speechrecognition
python-jose
passlib